fix(): reorganizes apps to break cyclic dependencies
