--- mockserver/authentication/models.py
@@ -1,14 +1,0 @@
-from django.db import models
-
-from base.models import DateAwareModel
-from tenants.models import Tenant
-
-
-class TenantAwareModel(DateAwareModel):
-    tenant = models.ForeignKey(
-        Tenant,
-        on_delete=models.CASCADE
-    )
-
-    class Meta:
-        abstract = True
