+++ mockserver/mockserver/settings.py
@@ -44,7 +44,7 @@ INSTALLED_APPS = [
     'rest_framework',
     'corsheaders',
 
-    'authentication',
+    'api',
     'mocks',
     'tenants',
 ]
